… src/Hosting/Hosting/test/HostingApplicationDiagnosticsTests.cs