Current behavior:

1. if (loggingEnabled || diagnosticsListenerEnabled), new up Activity
2. activity.SetParentId(headers["traceparent"] or headers["request-id"])
3. activity Add Baggage from headers["baggage"] or headers["Correlation-Context"]
4. _diagnosticsListener.OnActivityImport(activity, httpContext)
4. activity.Start()
5. _diagnosticsListener.Write(ActivityStartKey, httpContext)

New behavior:

1. if (loggingEnabled || diagnosticsListenerEnabled), new up a dummyListener
2. Defer creation of Activity to ActivitySource

if (activity is not null)
{

  3. Add baggage
  4. _diagnosticsListener.OnActivityImport(activity, httpContext)
  5. _diagnosticListener.Write(ActivityStartKey, httpContext);

}


Questions:
1. Baggage isn't present on Activity before starting
2. Is it okay if we don't fire diagnosticsListener.OnActivityImport? Is ordering important? Can I do it after Activity.Start()?