@shirhatti 2021-02-23
Before: `\crank.exe --config https://raw.githubusercontent.com/aspnet/Benchmarks/master/scenarios/json.benchmarks.yml --scenario json --profile aspnet-perf-lin --application.environmentVariables ASPNETCORE_LogLevel=Error --application.framework net6.0 --application.options.displayOutput true`

```
| application           |                           |
| --------------------- | ------------------------- |
| CPU Usage (%)         | 97                        |
| Cores usage (%)       | 1,164                     |
| Working Set (MB)      | 180                       |
| Build Time (ms)       | 4,726                     |
| Start Time (ms)       | 207                       |
| Published Size (KB)   | 280,235                   |
| .NET Core SDK Version | 6.0.100-preview.3.21123.4 |


| load                   |           |
| ---------------------- | --------- |
| CPU Usage (%)          | 50        |
| Cores usage (%)        | 594       |
| Working Set (MB)       | 38        |
| Start Time (ms)        | 0         |
| First Request (ms)     | 127       |
| Requests/sec           | 380,244   |
| Requests               | 5,741,726 |
| Mean latency (ms)      | 0.84      |
| Max latency (ms)       | 52.46     |
| Bad responses          | 0         |
| Socket errors          | 0         |
| Read throughput (MB/s) | 55.12     |
| Latency 50th (ms)      | 0.65      |
| Latency 75th (ms)      | 0.73      |
| Latency 90th (ms)      | 0.82      |
| Latency 99th (ms)      | 8.60      |
```

After: `crank.exe --config https://raw.githubusercontent.com/aspnet/Benchmarks/master/scenarios/json.benchmarks.yml --scenario json --profile aspnet-perf-lin --application.environmentVariables ASPNETCORE_LogLevel=Error --application.options.outputFiles C:\Users\shirh\source\repos\aspnetcore\artifacts\bin\Microsoft.AspNetCore.Hosting\Release\net6.0\Microsoft.AspNetCore.Hosting.dll --application.framework net6.0 --application.options.displayOutput true`

```
| application           |                           |
| --------------------- | ------------------------- |
| CPU Usage (%)         | 97                        |
| Cores usage (%)       | 1,164                     |
| Working Set (MB)      | 179                       |
| Build Time (ms)       | 12,768                    |
| Start Time (ms)       | 217                       |
| Published Size (KB)   | 280,235                   |
| .NET Core SDK Version | 6.0.100-preview.3.21123.4 |


| load                   |           |
| ---------------------- | --------- |
| CPU Usage (%)          | 51        |
| Cores usage (%)        | 611       |
| Working Set (MB)       | 48        |
| Build Time (ms)        | 7,868     |
| Start Time (ms)        | 0         |
| Published Size (KB)    | 76,407    |
| .NET Core SDK Version  | 3.1.406   |
| First Request (ms)     | 126       |
| Requests/sec           | 397,795   |
| Requests               | 6,006,528 |
| Mean latency (ms)      | 0.78      |
| Max latency (ms)       | 40.68     |
| Bad responses          | 0         |
| Socket errors          | 0         |
| Read throughput (MB/s) | 57.66     |
| Latency 50th (ms)      | 0.62      |
| Latency 75th (ms)      | 0.70      |
| Latency 90th (ms)      | 0.79      |
| Latency 99th (ms)      | 7.72      |
```

@Tratcher @sebastienros  Is there a better benchmark? I remember y'all mentioning there is a YARP benchmark for which Activity creation in hosting was on the hot path?