@sebastienros 2021-02-23
In the scenarios page, look for yarp. Beware that the sample command is for httpclient though.