@shirhatti 2021-02-23
Before: `crank.exe --config https://raw.githubusercontent.com/aspnet/Benchmarks/master/scenarios/proxy.benchmarks.yml --scenario proxy-httpclient --profile aspnet-perf-lin --application.framework net6.0 --application.options.displayOutput true`

```
| downstream            |         |
| --------------------- | ------- |
| CPU Usage (%)         | 87      |
| Cores usage (%)       | 1,043   |
| Working Set (MB)      | 149     |
| Build Time (ms)       | 4,180   |
| Start Time (ms)       | 287     |
| Published Size (KB)   | 95,075  |
| .NET Core SDK Version | 3.1.406 |


| application           |                           |
| --------------------- | ------------------------- |
| CPU Usage (%)         | 95                        |
| Cores usage (%)       | 1,136                     |
| Working Set (MB)      | 189                       |
| Build Time (ms)       | 4,142                     |
| Start Time (ms)       | 160                       |
| Published Size (KB)   | 263,567                   |
| .NET Core SDK Version | 6.0.100-preview.3.21123.4 |


| load                   |           |
| ---------------------- | --------- |
| CPU Usage (%)          | 40        |
| Cores usage (%)        | 475       |
| Working Set (MB)       | 42        |
| Start Time (ms)        | 111       |
| First Request (ms)     | 213       |
| Requests               | 1,878,924 |
| Bad responses          | 0         |
| Mean latency (us)      | 2,038     |
| Max latency (us)       | 51,858    |
| Requests/sec           | 125,322   |
| Requests/sec (max)     | 136,375   |
| Read throughput (MB/s) | 14.45     |
```

After: `crank.exe --config https://raw.githubusercontent.com/aspnet/Benchmarks/master/scenarios/proxy.benchmarks.yml --scenario proxy-httpclient --profile aspnet-perf-lin --application.options.outputFiles C:\Users\shirh\source\repos\aspnetcore\artifacts\bin\Microsoft.AspNetCore.Hosting\Release\net6.0\Microsoft.AspNetCore.Hosting.dll --application.framework net6.0 --application.options.displayOutput true`

```
| downstream            |         |
| --------------------- | ------- |
| CPU Usage (%)         | 86      |
| Cores usage (%)       | 1,038   |
| Working Set (MB)      | 148     |
| Build Time (ms)       | 5,190   |
| Start Time (ms)       | 288     |
| Published Size (KB)   | 95,075  |
| .NET Core SDK Version | 3.1.406 |


| application           |                           |
| --------------------- | ------------------------- |
| CPU Usage (%)         | 95                        |
| Cores usage (%)       | 1,137                     |
| Working Set (MB)      | 186                       |
| Build Time (ms)       | 5,000                     |
| Start Time (ms)       | 170                       |
| Published Size (KB)   | 263,567                   |
| .NET Core SDK Version | 6.0.100-preview.3.21123.4 |


| load                   |           |
| ---------------------- | --------- |
| CPU Usage (%)          | 40        |
| Cores usage (%)        | 476       |
| Working Set (MB)       | 52        |
| Build Time (ms)        | 5,820     |
| Start Time (ms)        | 111       |
| Published Size (KB)    | 77,083    |
| .NET Core SDK Version  | 3.1.406   |
| First Request (ms)     | 207       |
| Requests               | 1,857,181 |
| Bad responses          | 0         |
| Mean latency (us)      | 2,062     |
| Max latency (us)       | 50,195    |
| Requests/sec           | 124,134   |
| Requests/sec (max)     | 138,900   |
| Read throughput (MB/s) | 14.29     |
```