@sebastienros 2021-02-23
`--scenario proxy-httpclient` this is not yarp, this is a generic httpclient proxy without logging. Use `--scenario proxy-yarp` intead