@sebastienros 2021-02-23
Another tip. You can save the results with `--output old.json`, and `--output new.json` then do `crank compare old.json new.json` and get a nice comparison table to share