@shirhatti 2021-02-23
Old:
```
crank.exe --config https://raw.githubusercontent.com/aspnet/Benchmarks/master/scenarios/proxy.benchmarks.yml --scenario proxy-yarp --profile aspnet-perf-lin --application.framework net6.0 --output old.json
```

New:
```
crank.exe --config https://raw.githubusercontent.com/aspnet/Benchmarks/master/scenarios/proxy.benchmarks.yml --scenario proxy-yarp --profile aspnet-perf-lin --application.options.outputFiles C:\Users\shirh\source\repos\aspnetcore\artifacts\bin\Microsoft.AspNetCore.Hosting\Release\net6.0\Microsoft.AspNetCore.Hosting.dll --application.framework net6.0
```

```
| downstream            | old     | new     |         |
| --------------------- | ------- | ------- | ------- |
| CPU Usage (%)         |      85 |      82 |  -3.53% |
| Cores usage (%)       |   1,014 |     984 |  -2.96% |
| Working Set (MB)      |     149 |     149 |   0.00% |
| Build Time (ms)       |   5,219 |   4,243 | -18.70% |
| Start Time (ms)       |     295 |     293 |  -0.68% |
| Published Size (KB)   |  95,075 |  95,075 |   0.00% |
| .NET Core SDK Version | 3.1.406 | 3.1.406 |         |


| application           | old                       | new                       |         |
| --------------------- | ------------------------- | ------------------------- | ------- |
| CPU Usage (%)         |                        96 |                        97 |  +1.04% |
| Cores usage (%)       |                     1,154 |                     1,160 |  +0.52% |
| Working Set (MB)      |                       212 |                       212 |   0.00% |
| Build Time (ms)       |                     7,003 |                     5,355 | -23.53% |
| Start Time (ms)       |                       367 |                       352 |  -4.09% |
| Published Size (KB)   |                   263,844 |                   263,844 |   0.00% |
| .NET Core SDK Version | 6.0.100-preview.3.21123.4 | 6.0.100-preview.3.21123.4 |         |


| load                   | old       | new       |         |
| ---------------------- | --------- | --------- | ------- |
| CPU Usage (%)          |        35 |        35 |   0.00% |
| Cores usage (%)        |       417 |       417 |   0.00% |
| Working Set (MB)       |        42 |        42 |   0.00% |
| Start Time (ms)        |       111 |       110 |  -0.90% |
| First Request (ms)     |       246 |       239 |  -2.85% |
| Requests               | 1,443,811 | 1,426,715 |  -1.18% |
| Bad responses          |         0 |         0 |         |
| Mean latency (us)      |     2,653 |     2,683 |  +1.13% |
| Max latency (us)       |    53,336 |    72,823 | +36.54% |
| Requests/sec           |    96,441 |    95,419 |  -1.06% |
| Requests/sec (max)     |   105,120 |   104,774 |  -0.33% |
| Read throughput (MB/s) |     11.11 |     10.98 |  -1.18% |
```